+++ frontend/src/services/index.ts
@@ -1,0 +1,5 @@
+import generateResponse from "./chatServices/generateResponse";
+
+export {
+    generateResponse
+}
