+++ frontend/src/types/chatType.ts
@@ -1,0 +1,10 @@
+type Message = {
+    id: string;
+    content: string;
+    sender: "user" | "assistant";
+    timestamp: Date;
+}
+
+export type {
+    Message
+}
