+++ frontend/src/types/index.ts
@@ -1,0 +1,7 @@
+import { DocLink, TechDoc } from "./docType";
+import { Message } from "./chatType";
+
+export type {
+    TechDoc, DocLink, 
+    Message
+}
